More tweaks to grid collisions. This should fix bad swaps once and for all.
The issue was caused by two factors:

1. When multiple collisions occur, they are processed one by one. If the item collided with
is large enough, the first item could effectively leapfrog the second, causing a swap. I
fixed this by only moving the item one unit at a time.

2. When a small item collides with a larger, it is nice to have the swap optimization
that moves the collision above, so that the user doesn't have to drag all the way below
the collision (and far past the target) to trigger it. However, if this happens
during a large moving cascade, it can also cause switches. I fixed this by only
enabling this behavior when colliding with the currently dragged/resized element.