(#889) change breakpoint/layout event order
In Responsive, the onBreakpointChange event occurs after the
onLayoutChange event. So if the onLayoutChange event is used to save
the layout based upon the current breakpoint, the save will overwrite
the previous layout becuase the breakpoint you would be updating will
be the previous layout before changing breakpoints.

This PR will swap the order to onBreakpointChange -> onLayoutChange