feat(render): Add shouldComponentUpdate to ReactGridLayout & GridItem.
This improves performance quite a bit in most cases, is implemented safely
to avoid breakage on equal arrays and objects.

To do so, while avoiding the maintenance burden of keeping a keylist
in sync with propTypes, we read propTypes using babel-plugin-preval
then construct a function to do the fastest possible comparison.

This is *a lot* faster than isEqual while avoiding unnecessary
comparison of e.g. `children`.