fix/print-layout-item-in-wrong-place
If there are statics in the layout, then all items could potentially need moving, so you can't break out of the item-moving code early. Try removing the implementation and see how the test fails.

A PR to the original repo with this solution is coming, but first want to test in our own application.