#1394 - Dynamic configuration of droppingItem
Idea here is to allow the `droppingItem` to be adjusted based on what is being dragged over.
Question - maybe we should prevent the `i` property from being overwritten (and limit it to `w` and `h`). I am not sure but dynamically changing `i` may cause issues for removing the placeholder. I don't see any use case for needing to change `i` anyway.