feat(callbacks): add `ignoreClickOnly` prop.
If this is `true`, `onDragStart` is not called until an actual drag occurs.

This makes it easier to listen to `onDragStart` because you don't
have to ignore simple clicks.
As this is a breaking change, it will be behind this prop until 2.0.0,
after which **it will become the default behavior**.