Fix dragover positioning when using CSS translate
When using CSS transforms to translate the grid the calculated dragover
positions was offset within the grid.

The issue is layerX and layerY are not relative to the translation.
Also they are not standard according to MDN.

This commit changes approach to measure the final on-screen position of
the grid and compare that with clientX and clientY from the event.