Avoid possible stack overflows into `resolveCompactionCollision`
While running `react-grid-layout` on one of my project, I encountered a case leading `resolveCompactionCollision` into a stack overflow. The reason being that the cursion had to apply on too many colliding items and thus led us to a runtime crash.

The current PR changes the implementation of the algorithm into a non recursive version by replacing the native stack handled by the browser (and pretty limited in size) my a manually handled one backed by an array and behaving exactly the same way.